♻️ Refactoring the Challenges page for a smaller componenets
splitting the code to make it easier to read
