+++ src/interfaces/index.ts
@@ -1,2 +1,3 @@
 export type { PersonalInfo, Skill } from './personalInfo.interfaces';
 export type { Nullable } from './common';
+export type { CodeChallenge, TestResults } from './codeChallenges.interfaces';
